Separate tiles and tile colors in libhex, make more gc-friendly
Previously, libhex functions which returned a tile typically returned a
pointer to a hex_color, which isn't very friendly for alternative
implementations. There is now a separate hex_tile struct.

Additionally, spitting out pointers into the guts of the object isn't
especially friendly if the hex_board is stored in gc-controlled memory,
where it may be moved around. As an alternative, many functions now have
variants which accept board indices. A side-effect of this is that the
API no longer specifies that the internal representation of the board is
in column-major order.

"Unsafe" versions of several functions, which don't do bounds checking
and don't return an error code, have been added for cowboys/language
bindings. The latter tend to call multiple functions, resulting in
pointless bounds-checking on values that were just returned from another
function on the same board.

Several functions have been renamed to make them more consistent.