Add a libhex routine for flipping boards
i.e. exchanging the axes and colors. Being able to flip the board makes
it easier for AIs, since we can just write all the code assuming we're a
particular color, and if we're not that color, flip the board at the
start of the move, then flip our move back at the end.

The dummy AI has been tweaked to use the new routine.