New AI: "disruptor"
It's not super brilliant--I can beat it without too much effort--but it
plays rather better than dummy. It only does one move of lookahead, and
picks a move that maximises its opponents path cost and minimises its
own, preferring the former if there isn't one that does both.

Also, it's slow. It can take several seconds to pick a move on 11-by-11
boards during the early game, though it speeds up once some tiles have
been placed, since it'll be able to rule out a lot of tiles quickly.

It's written in Chicken Scheme, so you need that on your system. The
CMake code for building it makes a bunch of horrible nonportable
assumptions, namely that the compiler is in your path. I'll come back
and write a proper module later.