Improve first-move strategy for "disruptor"
Instead of painstakingly contemplating every single tile on the board
for the first move (which takes a while on larger boards) before just
picking one randomly, pick from squares near the middle. This restricts
the search space for later moves enough that it's playable on a 14x14
board if you're patient.