Make hextk DPI-aware under Tclkit
The Tclkit we're using isn't marked as DPI-aware, and attempting to add
a manifest using mt.exe apparently stripped the extra data which makes
it work properly.

We get around this problem by exposing SetProcessDPIAwarenessContext and
GetDpiForWindow to Tcl, which allows us to tell Windows that we know
what we're doing and then fix up the Tk scaling ourselves.

This hasn't been tested on a non-high-DPI system, nor has it been tested
on Windows 7, but it should probably work given that the relevant
procedures are loaded dynamically and sane default values have been set.